Added message if Git was not found.
A few custom library paths have been removed, which were used to
provide platform specific libraries for the build. We now expect
users to install dependencies on all platforms via their systems
means.

Also, if Git was not found and thus revision information could
not be extracted, we print out a message while configuring the
build.