Prepare source grouping for Win32 builds.
Source and header files can be grouped for Visual Studio solutions.
Currently these are just split into headers and sources. This macro
allows us to create child groups below sources and headers based on
a given topic.

An example would be:

    mangos_source_group_topic(${target_sources} PacketHandlers)

This would add headers and sources below into the folders

* Header Files\PacketHandlers
* Source Files\PacketHandlers

The intent is to provide a more readable project for larger sources.